+++ src/commands/index.ts
@@ -125,6 +125,7 @@ import { deleteWarn } from './moderation/warn/deleteWarn';
 import { atc } from './a32nx/atc';
 import { market } from './support/market';
 import { takeoffIssues } from './a32nx/takeoffissues';
+import { dfd } from './general/dfd';
 
 const commands: CommandDefinition[] = [
     ping,
@@ -252,6 +253,7 @@ const commands: CommandDefinition[] = [
     atc,
     market,
     takeoffIssues,
+    dfd,
 ];
 
 const commandsObject: { [k: string]: CommandDefinition } = {};